# openApiDocumentation.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.1.0",
    "title": "Musea API",
    "description": "Musea REST API for retrieving all the information about our museums, expositions and artworks",
    "contact": {
      "name": "Musea Team",
      "email": "info@musea.com",
      "url": "https://www.musea.com"
    }
  },
...
  "paths": {
    "/museums": {
      "get": {
        "tags": ["Museums"],
        "summary": "Get all museums",
        "description": "Retrieving all the museums",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Museums were retrieved successfully",
            "content": {
              "application/json": {
              }
            }
          },
          "404": {
            "description": "No museums found",
            "content": {
              "application/json": {
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Museums"],
        "summary": "Create a new museum",
        "description": "Creating a new museum with the given info",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "description": "Name of the new museum"
          },
          {
            "name": "address",
            "in": "query",
            "required": true,
            "description": "Address of the new museum"
          },
          {
            "name": "city",
            "in": "query",
            "required": true,
            "description": "City of the new museum"
          },
          {
            "name": "country",
            "in": "query",
            "required": true,
            "description": "Country of the new museum"
          },
          {
            "name": "ca",
            "in": "query",
            "required": true,
            "description": "Catalan description of the new museum"
          },
          {
            "name": "es",
            "in": "query",
            "required": true,
            "description": "Spanish description of the new museum"
          },
          {
            "name": "en",
            "in": "query",
            "required": true,
            "description": "English description of the new museum"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "restrictions": {
                    "type": "object",
                    "properties": {
                      "ca": {
                        "type": "string",
                        "example": "Restricció en català"
                      },
                      "es": {
                        "type": "string",
                        "example": "Restricción en español"
                      },
                      "en": {
                        "type": "string",
                        "example": "Restriction in english"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The museum was created successfully",
            "content": {
              "application/json": {
              }
            }
          }
        }
      }
    },
    "/museums/{museumId}": {
      "get": {
        "tags": ["Museums"],
        "summary": "Get museum by id",
        "description": "Retrieving a museum by the given id",
        "parameters": [
          {
            "name": "museumId",
            "in": "path",
            "required": true,
            "description": "Id of the desired museum"
          }
        ],
        "responses": {
          "200": {
            "description": "The museum was retrieved",
            "content": {
              "application/json": {
              }
            }
          },
          "404": {
            "description": "There is no museum for such id",
            "content": {
              "application/json": {
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Museums"],
        "summary": "Create a new exposition",
        "description": "Creating a new exposition with the given info",
        "parameters": [
          {
            "name": "museumId",
            "in": "path",
            "required": true,
            "description": "Id of the hosting museum"
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "description": "Name of the new exposition"
          },
          {
            "name": "room",
            "in": "query",
            "required": true,
            "description": "Room of the new exposition"
          },
          {
            "name": "ca",
            "in": "query",
            "required": true,
            "description": "Catalan description of the new exposition"
          },
          {
            "name": "es",
            "in": "query",
            "required": true,
            "description": "Spanish description of the new exposition"
          },
          {
            "name": "en",
            "in": "query",
            "required": true,
            "description": "English description of the new exposition"
          }
        ],
        "responses": {
          "200": {
            "description": "The exposition was created successfully",
            "content": {
              "application/json": {
              }
            }
          },
          "500": {
            "description": "Could not update the museum with the new exposition",
            "content": {
              "application/json": {
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Museums"],
        "summary": "Delete museum",
        "description": "Deleting a museum by the given id",
        "parameters": [
          {
            "name": "museumId",
            "in": "path",
            "required": true,
            "description": "Id of the desired museum"
          }
        ],
        "responses": {
          "200": {
            "description": "The museum was deleted",
            "content": {
              "application/json": {
              }
            }
          },
          "404": {
            "description": "There is no museum for such id",
            "content": {
              "application/json": {
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Museums"],
        "summary": "Edit museum",
        "description": "Editing a museum by the given id",
        "parameters": [
          {
            "name": "museumId",
            "in": "path",
            "required": true,
            "description": "Id of the museum"
          },
          {
            "name": "ca",
            "in": "query",
            "required": false,
            "description": "Catalan description of the museum"
          },
          {
            "name": "es",
            "in": "query",
            "required": false,
            "description": "Spanish description of the museum"
          },
          {
            "name": "en",
            "in": "query",
            "required": false,
            "description": "English description of the museum"
          },
          {
            "name": "image",
            "in": "query",
            "required": false,
            "description": "Url of the museum's image"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "restrictions": {
                    "type": "object",
                    "properties": {
                      "ca": {
                        "type": "string",
                        "example": "Restricció en català"
                      },
                      "es": {
                        "type": "string",
                        "example": "Restricción en español"
                      },
                      "en": {
                        "type": "string",
                        "example": "Restriction in english"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The museum was edited successfully",
            "content": {
              "application/json": {
              }
            }
          },
          "404": {
            "description": "There is no museum for such id",
            "content": {
              "application/json": {
              }
            }
          }
        }
      }
    },
...
    "/info": {
      "get": {
        "tags": ["Museums"],
        "summary": "Get museum information",
        "description": "Retrieving the museum's info: opening hours, restrictions, etc.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "description": "Name of the desired museum"
          },
          {
            "name": "city",
            "in": "query",
            "required": true,
            "description": "City of the desired exposition"
          }
        ],
        "responses": {
          "200": {
            "description": "The museum information was retrieved",
            "content": {
              "application/json": {
              }
            }
          },
          "500": {
            "description": "External API server error",
            "content": {
              "application/json": {
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "description": "Retrieving all the users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users were retrieved successfully",
            "content": {
              "application/json": {
              }
            }
          },
          "404": {
            "description": "No users found",
            "content": {
              "application/json": {
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user",
        "description": "Creating a new user with the given info",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "description": "Username of the new user"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "description": "Email of the new user"
          }
        ],
        "responses": {
          "200": {
            "description": "The user was created successfully",
            "content": {
              "application/json": {
              }
            }
          },
          "500": {
            "description": "Could not create the new user",
            "content": {
              "application/json": {
              }
            }
          }
        }
      }
    },
...
  }
}